test(configuration): group CLI source contracts
